# alzawaj-project-backend/src/config/openapi.yaml
openapi: 3.0.0
info:
  title: Zawag Islamic Marriage Platform API
  version: 1.0.0
  description: API documentation for the Zawag Islamic Marriage Platform
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://alzawaj-backend.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MarriageRequest:
      type: object
      properties:
        id:
          type: string
          example: "6051d2d7e1b8a20015d23f5c"
        senderId:
          type: string
          example: "6051d2d7e1b8a20015d23f5a"
        receiverId:
          type: string
          example: "6051d2d7e1b8a20015d23f5d"
        status:
          type: string
          enum: [pending, accepted, rejected]
          example: "pending"
        message:
          type: string
          example: "Assalamu alaikum, I am interested in getting to know you better inshallah."
        createdAt:
          type: string
          format: date-time
          example: "2021-03-17T10:30:00Z"
        expiresAt:
          type: string
          format: date-time
          example: "2021-04-17T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "بيانات التسجيل غير صحيحة"
        statusCode:
          type: integer
          example: 400
        error:
          type: array
          items:
            type: string
          example:
            - "تأكيد كلمة المرور مطلوب"
            - "الاسم يجب أن يكون بين 2-100 حرف"
            - "العمر يجب أن يكون بين 18-100 سنة"
            - "حالة اللحية مطلوبة للذكور"
            - "الحالة المالية مطلوبة للذكور"
            - "نوع السكن مطلوب للذكور"
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/requests:
    get:
      tags: [Admin]
      summary: Get marriage requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      requests:
                        type: array
                        items:
                          $ref: '#/components/schemas/MarriageRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/reports:
    get:
      tags: [Admin]
      summary: Get reports
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reports:
                        type: array
                        items:
                          type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: Get admin settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Admin]
      summary: Update admin settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageLimits:
                  type: object
                chatSettings:
                  type: object
                moderationSettings:
                  type: object
                themeSettings:
                  type: object
      responses:
        '200':
          description: Settings successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...